Decouple 'AtmosphereParameters' from SystemBody, thus from graphics* and terrain*
