+++ src/galaxy/AtmosphereParameters.h
@@ -1,0 +1,13 @@
+#ifndef ATMOSPHEREPARAMETERS_H_INCLUDED
+#define ATMOSPHEREPARAMETERS_H_INCLUDED
+
+struct AtmosphereParameters {
+	float atmosRadius;
+	float atmosInvScaleHeight;
+	float atmosDensity;
+	float planetRadius;
+	Color atmosCol;
+	vector3d center;
+};
+
+#endif // ATMOSPHEREPARAMETERS_H_INCLUDED
